mpi_cfd2.f90
This script computes the Laplacian of the variable u using second-order accurate central finite differences on Np number of processors. The domain is divided up in the x direction, so the number of cells in x (Nx) must be divisible by the number of MPI ranks (Np). The is one layer of halo cells surrounding the domain and the domain chunks for parallel computation.